src/entertainment_education.py
```python
import random
from src.data.ai_services_data import aiServices
from src.utils.helpers import displayContent

class EntertainmentEducation:
    def __init__(self, userProfile):
        self.userProfile = userProfile

    def provideContent(self):
        # Fetch content based on user's preferences
        content = self._fetchContentBasedOnPreferences()
        
        # Display the content
        displayContent(content)

    def _fetchContentBasedOnPreferences(self):
        # Fetch AI services related to entertainment and education
        entertainmentEducationServices = [service for service in aiServices if service['category'] in ['Entertainment', 'Education']]
        
        # Filter services based on user's preferences
        preferredServices = [service for service in entertainmentEducationServices if service['name'] in self.userProfile['preferences']]
        
        # If no preferred services, return a random service
        if not preferredServices:
            return random.choice(entertainmentEducationServices)
        
        # Else, return a random preferred service
        return random.choice(preferredServices)
```