src/tests/test_entertainment_education.py
```python
import unittest
from src.entertainment_education import provideContent

class TestEntertainmentEducation(unittest.TestCase):

    def setUp(self):
        self.userProfile = {
            "name": "John Doe",
            "age": 25,
            "preferences": {
                "games": ["Spyro The Dragon", "Crash Bandicoot"],
                "education": ["AI", "Machine Learning"]
            }
        }

    def test_provideContent(self):
        content = provideContent(self.userProfile)
        self.assertIsNotNone(content)
        self.assertIn('content', content)
        self.assertIn('type', content)
        self.assertIn(content['type'], ['game', 'education'])

    def test_provideContent_no_preferences(self):
        self.userProfile['preferences'] = {}
        content = provideContent(self.userProfile)
        self.assertIsNotNone(content)
        self.assertIn('content', content)
        self.assertIn('type', content)
        self.assertIn(content['type'], ['game', 'education'])

if __name__ == '__main__':
    unittest.main()
```